bugfix: crash on Android stop command
After the system stop command, the callback in hwc3 is destroyed.

After hwc3 calls UnregisterVsyncCallback to destroy, it needs to notify
vsyncwork in time and prevent it from continuing to call the hwc3 callback.

Therefore, the timing of the callback variable update in vsync_worker in
hwc2 needs to be adjusted.

Change-Id: I6830815a3f1c815d065fada20d0b1bd2cc1c1986
Signed-off-by: Li Huang <putin.li@rock-chips.com>