BugFix : Fix the problem of UBoot logo residue.
The dynamic assign of DrmPlane for the RK3576 needs to consider
the DrmPlane configuration of the underlying UBoot logo.

If it differs from the HWC's expected UBoot configuration,
dynamic switching logic must be executed.

The current issue is that the dynamic switching logic requires
accurate DisplayId information. If it is not configured,
it may cause the dynamic switching to fail, resulting in the
logo residue problem.

RedmineId : Defect #491219

Change-Id: Iebe7979f377736e86b0a69a4a1def2a192b93b88
Signed-off-by: libin <bin.li@rock-chips.com>